refactor: modularize MCP tool definitions and registration logic to improve maintainability
